backend/requirements.txt
# Core
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-decouple==3.8
celery==5.3.1
redis==4.6.0

# Database
psycopg2-binary==2.9.7
django-redis==5.3.0

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.54.0

# WebSocket
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# WhatsApp Integration
twilio==8.5.0
requests==2.31.0

# NLP & ML
spacy==3.6.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.25.2

# API Documentation
drf-spectacular==0.26.4

# Testing
pytest==7.4.0
pytest-django==4.5.2
factory-boy==3.3.0

# Monitoring
sentry-sdk==1.29.2

# Utils
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3